+++ CovidWatch/Server/LocalGoogleExposureNotificationServer.swift
@@ -1,0 +1,11 @@
+//
+//  Created by Zsombor Szabo on 23/05/2020.
+//  
+//
+
+import Foundation
+
+// TODO
+@available(iOS 13.5, *)
+public class LocalGoogleExposureNotificationServer: GoogleExposureNotificationServer {
+}
